🚸 Improve AddBookPage to server component
